Title: 2개씩 합치기, Time: 544ms, Memory: 21MB, Status: Passed - CodeTree

+++ 240724/2개씩 합치기/combine-two.java
@@ -17,7 +17,7 @@ public class Main {
             pq.add(num);
         }
 
-        int cost = 0;
+        long cost = 0;
         while (pq.size() > 1) {
             int num1 = pq.remove();
             int num2 = pq.remove();
